requirements.txt
asgiref==3.3.4
Django==2.0.13
django-email-extras==0.3.4
django-forms-builder==0.14.0
future==0.15.0
Pillow==8.2.0
python-gnupg==0.4.7
pytz==2021.1
sphinx-me==0.3
sqlparse==0.4.1
Unidecode==1.2.0
